SUDO-125 SUDO-157 finished edit of users and fine tuned validation for editing and deleting users

+++ CutePuppies/src/main/java/com/sg/cutepuppies/controllers/UserController.java
@@ -9,12 +9,16 @@ import com.sg.cutepuppies.daos.UserDaoInterface;
 import com.sg.cutepuppies.models.User;
 import java.util.List;
 import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
+import org.springframework.http.HttpRequest;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -44,6 +48,18 @@ public class UserController {
         List<User> users = userDao.getAllUsers();
         model.addAttribute("formUser", new User());
         model.addAttribute("users", users);
+        model.addAttribute("addEdit", "add");
+        return "manage-users";
+    }
+    
+    @RequestMapping(value = {"/edit/{userId}"}, method = RequestMethod.GET)
+    public String editUser(Model model, @PathVariable int userId) {
+        List<User> users = userDao.getAllUsers();
+        User user = userDao.getUserById(userId);
+        user.setUserPassword("");
+        model.addAttribute("formUser", user);
+        model.addAttribute("users", users);
+        model.addAttribute("addEdit", "edit");
         return "manage-users";
     }
 
@@ -68,6 +84,7 @@ public class UserController {
             model.addAttribute("userPasswordEmpty", userPasswordEmpty);
             model.addAttribute("userRoleCodeEmpty", userRoleCodeEmpty);
             model.addAttribute("userPasswordsMismatch", userPasswordsMismatch);
+            model.addAttribute("addEdit", "add");
             return "manage-users";
         }
         //  - Hash the password and then set it on the User object before saving it
@@ -79,15 +96,58 @@ public class UserController {
         return "redirect:userList";
     }
 
-    @RequestMapping(value = "/update", method = RequestMethod.PUT)
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateUser(@Valid @RequestBody User user) {
+    @RequestMapping(value = "/edit", method = RequestMethod.POST)
+    public String updateUser(@Valid @ModelAttribute User user, Model model, String password2) {
+        boolean userNameExists = (userDao.getUserIdByUsername(user.getUserName()) != 0 
+                && userDao.getUserIdByUsername(user.getUserName()) != user.getUserId());
+        boolean userEmailExists = (userDao.getUserIdByEmail(user.getUserEmail()) != 0 
+                && userDao.getUserIdByEmail(user.getUserEmail()) != user.getUserId());
+        boolean userNameEmpty = (user.getUserName() == "" || user.getUserName() == null);
+        boolean userEmailEmpty = (user.getUserEmail()== "" || user.getUserEmail()== null);
+        boolean userPasswordEmpty = (user.getUserPassword()== "" || user.getUserPassword()== null);
+        boolean userRoleCodeEmpty = (user.getRoleCode() == "" || user.getRoleCode() == null);
+        boolean userPasswordsMismatch = (!user.getUserPassword().equals(password2));
+        
+        if (userNameExists || userEmailExists || userNameEmpty || userEmailEmpty || userPasswordEmpty || userRoleCodeEmpty) {
+            List<User> users = userDao.getAllUsers();
+            user.setUserPassword("");
+            model.addAttribute("formUser", user);
+            model.addAttribute("users", users);
+            model.addAttribute("userNameExists", userNameExists);
+            model.addAttribute("userEmailExists", userEmailExists);
+            model.addAttribute("userNameEmpty", userNameEmpty);
+            model.addAttribute("userEmailEmpty", userEmailEmpty);
+            model.addAttribute("userPasswordEmpty", userPasswordEmpty);
+            model.addAttribute("userRoleCodeEmpty", userRoleCodeEmpty);
+            model.addAttribute("userPasswordsMismatch", userPasswordsMismatch);
+            model.addAttribute("addEdit", "edit");
+            return "manage-users";
+        }
+        //  - Hash the password and then set it on the User object before saving it
+        String clearPw = user.getUserPassword();
+        String hashPw = encoder.encode(clearPw);
+        user.setUserPassword(hashPw);
         userDao.updateUser(user);
+        
+        return "redirect:userList";
+        
     }
 
-    @RequestMapping(value = "/remove/{userId}", method = RequestMethod.GET)
-    public void removeUser(@Valid @RequestParam("userId") int userId) {
+    @RequestMapping(value = "/remove", method = RequestMethod.POST)
+    public String removeUser(int userId, Model model) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        int currentUserId = userDao.getUserIdByUsername(username);
+        if (currentUserId == userId) {
+            List<User> users = userDao.getAllUsers();
+            model.addAttribute("formUser", new User());
+            model.addAttribute("users", users);
+            model.addAttribute("isCurrentUser", true);
+            model.addAttribute("userId", userId);
+            model.addAttribute("addEdit", "add");
+            return "manage-users";
+        }
         userDao.deleteUser(userId);
+        return "redirect:userList";
     }
 
 }
